Create highest times value of an array
/*
In statistics, the mode of a group of numbers is the one that occurs the most often. For
example, given the list 1, 2, 3, 6, 4, 7, 5, 4, 6, 9, 4, the mode is 4, because it occurs three
[5]
Page 3 of 5
times. Write a C program that allows the user to enter a list of 15 numbers and then finds
and displays the mode.
Sample Input:
Enter 15 numbers: 4 1 5 3 2 4 3 6 0 9 3 7 6 3 8
Sample Output:
Mode is: 3

    
    */